Unify by-name
