+++ components/communes/index.js
@@ -8,7 +8,7 @@ import Summary from '../summary'
 import TechnicalDoc from '../technical-doc'
 
 import ByCode from './examples/by-code'
-import ByName from './examples/by-name'
+import ByCommuneName from './examples/by-commune-name'
 import CommunesList from './examples/communes-list'
 import ByLatLon from './examples/by-lat-lon'
 import AdvancedSearch from './examples/advanced-search'
@@ -30,7 +30,7 @@ const Communes = () => (
     <TechnicalDoc {...doc} />
 
     <ByCode {...examples[0]} />
-    <ByName {...examples[1]} />
+    <ByCommuneName {...examples[1]} />
     <CommunesList {...examples[2]} />
     <ByLatLon {...examples[3]} />
     <AdvancedSearch {...examples[4]} />
